add compound and other information for abstract. Run in DAG.

+++ pipeline/process_abstract.py
@@ -4,18 +4,30 @@ from langchain_groq import ChatGroq
 from langchain_mistralai import ChatMistralAI
 from langchain.prompts import ChatPromptTemplate
 import pickle
-from defs.abstract_class import Abstract, tcga_to_fullname
+from defs.abstract_class import Abstract
 import json
 from airflow.decorators import task
 import pandas as pd
 import os
 from .config import NUM_PARTITIONS, DEV_PAGE_LIMIT, STORAGE_DIR, ENVIRONMENT
+from pydantic import BaseModel, Field
+from typing import List, Optional
+from utils.vectorstores_ontology import VectorStore
 
+from utils.get_disease_terms import prepare_disease_synonyms
+from utils.get_gene_synonyms import prepare_gene_synonyms
+from helpers import log_progress
 from dotenv import load_dotenv
 load_dotenv()
 
-from pydantic import BaseModel, Field
-from typing import List, Optional
+
+# Prepare disease and gene synonyms
+disease_synonyms = prepare_disease_synonyms()
+gene_synonyms = prepare_gene_synonyms()
+
+vectorstore = VectorStore()
+vectorstore.prepare_lookups(gene_synonyms, disease_synonyms)
+vectorstore.prepare_vectorstores()
 
 def extract_abstract_info(abstract_text, model = 'mistral'):
     
@@ -32,11 +44,13 @@ def extract_abstract_info(abstract_text, model = 'mistral'):
             max_retries=2
         )
 
+
+    gene_context, disease_context = vectorstore.rag(abstract_text)
     
-    valid_full_names = list(tcga_to_fullname.values())
+    print(gene_context, disease_context)
     
     prompt = ChatPromptTemplate.from_messages([
-        ("system", f"""You are an expert at extracting information from academic abstracts, with a focus on genetics and oncology. Your task is to accurately and completely extract the requested information, paying close attention to details. Extract the information according to the following structure:
+    ("system", f"""You are an expert at extracting information from academic abstracts, with a focus on genetics and oncology. Your task is to accurately and completely extract the requested information, paying close attention to details. Extract the information according to the following structure:
 
 1. abstract_number: Extract ONLY the 4-digit number from the abstract number. Ignore any preceding or following characters. For example, if you see '0009AFNT-212', extract ONLY '0009'. If no 4-digit number is present, use 'n/a'.
 
@@ -46,23 +60,29 @@ def extract_abstract_info(abstract_text, model = 'mistral'):
    - name: The full name of the author, exactly as it appears in the abstract
    - affiliation: The institution or organization the author is affiliated with. If not available, use null.
 
-4. disease: A list of specific cancer indications or types mentioned in the abstract. Use ONLY the following full names: {', '.join(valid_full_names)}. If a cancer type is mentioned that's not in this list, do not include it. If no valid cancer types are mentioned, return an empty list.
+4. disease: A list of specific cancer indications or types mentioned in the abstract. Use ONLY the terms provided in the following disease context: {disease_context}. Map synonyms to the primary term. If it is about cancer but not a specific one, put 'Cancer general'. If a cancer type is mentioned that's not in this list, do not include it. If no valid cancer types are mentioned, return an empty list.
 
-5. gene: A list of genes mentioned in the abstract, using only official HUGO gene symbols. Include only genes directly related to the discussed cancer or research topic. If no genes are mentioned, return an empty list.
+5. gene: A list of genes mentioned in the abstract, using only the gene symbols provided in the following gene context: {gene_context}. Map synonyms to the primary term. Include only genes directly related to the discussed cancer or research topic. If no genes from this list are mentioned, return an empty list.
+
+6. organism: The type of organism for which the evidence is provided. Can be 'cell line', 'PDX', 'animal', 'human', or 'n/a' if not specified. Use 'animal' if any of the following are mentioned: mouse, mice, rat, rabbit, guinea pig, hamster, dog, pig, monkey, or primate. If multiple organism types are used, list all that apply.
+
+7. trial_stage: The stage of the trial or study. Can be 'preclinical', 'Phase I', 'Phase II', 'Phase III', 'post-approval', or 'n/a' if not applicable or not specified.
+
+8. compound_name: The name of the compound or drug mentioned in the abstract. E.g. 'GS-P-328', 'IOMX-0675', or 'Pembrolizumab'
+
+9. abstract_category: Assign the abstract to one of the following categories: 'Gene-Disease Associations' if it is about the impact of some genes on the disease, 'Diagnostics' if it is about methods for diagnosis and does not provide a novel finding on a gene, or 'Other'. Choose the most appropriate category based on the main focus of the abstract.
 
 Follow these guidelines strictly:
 - For the abstract number, extract ONLY the 4-digit number. Ignore any other characters.
-- For genes, use only official HUGO gene symbols. If a gene is mentioned using a different nomenclature, convert it to the HUGO symbol if you are certain of the correct conversion. If unsure, do not include the gene.
-- For cancer indications, use ONLY the full names provided in the list above. Do not include any cancer types not in this list.
-- Preserve all formatting, capitalization, and special characters in the title and abstract text exactly as they appear in the original.
 - Include all authors mentioned in the abstract, even if their affiliation is not provided.
 - If any information is not available or cannot be determined with certainty, use null for optional fields or an empty list for list fields.
 - Never guess or infer information that is not explicitly stated in the abstract.
 - Do not attempt to fill in missing information based on context or general knowledge.
 
 Accuracy is paramount. It is better to omit information than to provide potentially incorrect data. Under no circumstances should you guess or make assumptions about any data."""),
-        ("human", "{text}")
-    ])
+    ("human", "{text}")
+])
+
     
     # Create the extraction chain
     extraction_chain = prompt | llm.with_structured_output(schema=Abstract)
@@ -72,18 +92,38 @@ Accuracy is paramount. It is better to omit information than to provide potentia
 
 
 @task
-def process_abstracts_partition(partition_file):
-    df = pd.read_pickle(partition_file)
-    processed_abstracts = []
-    for _, row in df.iterrows():
-        json_result = extract_abstract_info(row)
-        processed_abstracts.append(json.loads(json_result))
+def process_abstracts_partition(batch_pages: List[int], model = 'mistral', **kwargs):
+    ti = kwargs['ti']
+    pages_processed_dir = os.path.join(STORAGE_DIR, ENVIRONMENT, 'processed_pages')    
+    pages_parsed_dir = os.path.join(STORAGE_DIR, ENVIRONMENT, f'parsed_pages_{model}')    
+    os.makedirs(pages_parsed_dir, exist_ok=True)   
     
-    output_file = partition_file.replace('.pkl', '_processed.json')
-    with open(output_file, 'w') as f:
-        json.dump(processed_abstracts, f, indent=2)
-    
-    return output_file
+    for i in batch_pages:
+        
+        input_file = os.path.join(pages_processed_dir, f'page_{i+1}.json')
+        output_file = os.path.join(pages_parsed_dir, f'page_{i+1}.json')
+        
+        if os.path.exists(output_file):
+            log_progress(f"Skipping page {i+1} as it has already been processed.")
+            continue
+        
+        log_progress(f"Processing page {i+1} with model {model}...")
+        with open(input_file) as f:
+            abstract = json.load(f)
+        abstract_text = abstract[0].get('content')
+        
+        # %%
+        result = extract_abstract_info(abstract_text, model='groq')
+        abstract_dict = result.dict()
+        abstract_dict['text'] = abstract_text
+        abstract_dict['page_number'] = abstract[0].get('page_number')
+        
+        json_result = json.dumps(abstract_dict, indent=2)        
+        
+        with open(output_file, 'w') as f:
+            f.write(json_result)
+        
+
 
 
 @task
